+++ crates/common/src/lib.rs
@@ -25,12 +25,14 @@ pub mod contracts;
 pub mod errors;
 pub mod evm;
 pub mod fs;
+pub mod mapping_slots;
 mod preprocessor;
 pub mod provider;
 pub mod reports;
 pub mod retry;
 pub mod selectors;
 pub mod serde_helpers;
+pub mod slot_identifier;
 pub mod term;
 pub mod traits;
 pub mod transactions;
